+++ build-tools-internal/src/main/java/org/elasticsearch/gradle/internal/InternalDistributionModuleCheckTaskProvider.java
@@ -54,8 +54,8 @@ public class InternalDistributionModuleCheckTaskProvider {
         "org.elasticsearch.geo",
         "org.elasticsearch.logging",
         "org.elasticsearch.lz4",
-        "org.elasticsearch.plugin.analysis.api",
-        "org.elasticsearch.plugin.api",
+        "org.elasticsearch.plugin.analysis",
+        "org.elasticsearch.plugin",
         "org.elasticsearch.pluginclassloader",
         "org.elasticsearch.securesm",
         "org.elasticsearch.server",
