+++ libs/plugin-api/src/main/java/module-info.java
@@ -6,7 +6,7 @@
  * Side Public License, v 1.
  */
 
-module org.elasticsearch.plugin.api {
+module org.elasticsearch.plugin {
     exports org.elasticsearch.plugin;
     exports org.elasticsearch.plugin.settings;
 }
