+++ libs/plugin-scanner/src/test/java/org/elasticsearch/plugin/scanner/NamedComponentScannerTests.java
@@ -64,13 +64,13 @@ public class NamedComponentScannerTests extends ESTestCase {
         Path jar = dirWithJar.resolve("plugin.jar");
         JarUtils.createJarWithEntries(jar, Map.of("p/A.class", InMemoryJavaCompiler.compile("p.A", """
             package p;
-            import org.elasticsearch.plugin.api.*;
+            import org.elasticsearch.plugin.*;
             import org.elasticsearch.plugin.scanner.test_model.*;
             @NamedComponent("a_component")
             public class A extends ExtensibleClass {}
             """), "p/B.class", InMemoryJavaCompiler.compile("p.B", """
             package p;
-            import org.elasticsearch.plugin.api.*;
+            import org.elasticsearch.plugin.*;
             import org.elasticsearch.plugin.scanner.test_model.*;
             @NamedComponent("b_component")
             public class B implements ExtensibleInterface{}
@@ -107,7 +107,7 @@ public class NamedComponentScannerTests extends ESTestCase {
             "p.CustomExtensibleInterface",
             """
                 package p;
-                import org.elasticsearch.plugin.api.*;
+                import org.elasticsearch.plugin.*;
                 import org.elasticsearch.plugin.scanner.test_model.*;
                 public interface CustomExtensibleInterface extends ExtensibleInterface {}
                 """,
@@ -115,14 +115,14 @@ public class NamedComponentScannerTests extends ESTestCase {
             "p.CustomExtensibleClass",
             """
                 package p;
-                import org.elasticsearch.plugin.api.*;
+                import org.elasticsearch.plugin.*;
                 import org.elasticsearch.plugin.scanner.test_model.*;
                 public class CustomExtensibleClass implements CustomExtensibleInterface {}
                 """,
             "p.A",
             """
                 package p;
-                import org.elasticsearch.plugin.api.*;
+                import org.elasticsearch.plugin.*;
                 import org.elasticsearch.plugin.scanner.test_model.*;
                 @NamedComponent("a_component")
                 public class A extends CustomExtensibleClass {}
@@ -130,7 +130,7 @@ public class NamedComponentScannerTests extends ESTestCase {
             "p.B",
             """
                 package p;
-                import org.elasticsearch.plugin.api.*;
+                import org.elasticsearch.plugin.*;
                 import org.elasticsearch.plugin.scanner.test_model.*;
                 @NamedComponent("b_component")
                 public class B implements CustomExtensibleInterface{}
