+++ server/src/main/java/org/elasticsearch/plugins/scanners/StablePluginsRegistry.java
@@ -27,7 +27,7 @@ public class StablePluginsRegistry {
     effectively means that an interface which extends another interface marked with @Extensible is also extensible
     NameToPluginInfo map is a map of Name to PluginInfo
     i.e.
-    org.elasticsearch.plugin.analysis.api.TokenFilterFactory ->
+    org.elasticsearch.plugin.analysis.TokenFilterFactory ->
         {"nori" -> {nori, org.elasticserach.plugin.analysis.new_nori.NoriReadingFormFilterFactory, classloaderInstance}
      */
     private final Map<String /*Extensible */, NameToPluginInfo> namedComponents;
